django-basic-tinymce-flatpages/requirements.txt
django>=1.6.0
django-tinymce
